move session to subfolder
