+++ python/sglang/srt/session/session_controller.py
@@ -24,7 +24,7 @@ from sglang.srt.managers.io_struct import (
     TokenizedGenerateReqInput,
 )
 from sglang.srt.managers.schedule_batch import FINISH_ABORT, Req
-from sglang.srt.mem_cache.session_aware_cache import SessionAwareCache
+from sglang.srt.session.session_aware_cache import SessionAwareCache
 from sglang.srt.utils.common import log_info_on_rank0
 
 if TYPE_CHECKING:
