+++ test/registered/unit/mem_cache/test_streaming_session_unit.py
@@ -5,7 +5,7 @@ import torch
 from sglang.srt.managers.schedule_batch import FINISH_ABORT
 from sglang.srt.mem_cache.base_prefix_cache import MatchResult
 from sglang.srt.mem_cache.common import release_kv_cache
-from sglang.srt.mem_cache.session_aware_cache import SessionAwareCache, SessionSlot
+from sglang.srt.session.session_aware_cache import SessionAwareCache, SessionSlot
 from sglang.test.ci.ci_register import register_cpu_ci
 
 register_cpu_ci(est_time=8, suite="stage-a-test-cpu")
